+++ includes/integration/integration.php
@@ -1,0 +1,20 @@
+<?php if ( ! defined( 'ABSPATH' ) ) exit;
+
+/**
+ * 3rd party plugins integrations
+ *
+ * @package     WebMan Amplifier
+ * @subpackage  Integration
+ * @copyright   WebMan Design, Oliver Juhas
+ *
+ * @since    1.3.22
+ * @version  1.3.22
+ */
+
+
+
+
+
+// Load integration files
+
+	require_once WMAMP_INCLUDES_DIR . 'integration/woosidebars/class-woosidebars.php';